requirements.txt
# OCR environment - Windows, CPU-only. Python 3.9+
# Install PyTorch CPU first, then the rest:
#   pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
#   pip install -r requirements.txt

# PyTorch (CPU) - install from index above
torch==2.10.0+cpu
torchvision==0.25.0+cpu
torchaudio==2.10.0+cpu

# Transformers (TrOCR)
transformers==5.0.0

# Image / preprocessing
opencv-python==4.13.0.90
pillow==12.0.0
numpy==2.3.5
matplotlib==3.10.8
scikit-image==0.26.0

# OCR
pytesseract==0.3.13
lmdb==1.4.0   # UTRNet read.py / dataset

# Layout (without Detectron2 - see SETUP_NOTES.md)
layoutparser==0.3.4

# Transitive / shared (pinned for reproducibility)
filelock==3.20.0
huggingface-hub==1.3.5
packaging==26.0
PyYAML==6.0.3
regex==2026.1.15
safetensors==0.7.0
tokenizers==0.22.2
typer-slim==0.21.1
typing-extensions==4.15.0
Jinja2==3.1.6
MarkupSafe==2.1.5
fsspec==2025.12.0
sympy==1.14.0
networkx==3.6.1
mpmath==1.3.0
contourpy==1.3.3
cycler==0.12.1
fonttools==4.61.1
kiwisolver==1.4.9
pyparsing==3.3.2
python-dateutil==2.9.0.post0
six==1.17.0
scipy==1.17.0
imageio==2.37.2
tifffile==2026.1.28
lazy-loader==0.4
pandas==3.0.0
iopath==0.1.10
pdfplumber==0.11.9
pdf2image==1.17.0
pdfminer.six==20251230
portalocker==3.2.0
tzdata==2025.3
anyio==4.12.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
cffi==2.0.0
cryptography==46.0.4
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
idna==3.11
tqdm==4.67.1
pycparser==3.0
pywin32==311
pypdfium2==5.3.0
shellingham==1.5.4
hf-xet==1.2.0
